requirements.txt
# Notion集成
notion-client>=2.2.1

# RSS处理
feedparser>=6.0.10

# HTTP请求
requests>=2.31.0
aiohttp>=3.9.0

# HTML解析
beautifulsoup4>=4.12.0

# 数据库
# SQLite是Python内置的，无需安装

# 翻译（DeepL或Google Translate）
# deep-translator>=1.11.0  # Google Translate（免费）
# deepl>=1.15.0            # DeepL（需要API密钥）

# Twitter API
# tweepy>=4.14.0           # Twitter API（需要API密钥）

# YouTube API
# google-api-python-client>=2.100.0

# 日期处理
python-dateutil>=2.8.2

# 环境变量管理
python-dotenv>=1.0.0

# 日志
# logging是Python内置的，无需安装

# 数据验证
pydantic>=2.5.0